+++ src/app/layout.tsx
@@ -4,8 +4,8 @@ import '@/styles/globals.css'
 import { Background } from '@/components/layout/background'
 import { Footer } from '@/components/layout/footer'
 import { Header } from '@/components/layout/header'
-import { buildPresenceScript } from '@/lib/achievements'
 import { buildThemeScript } from '@/lib/theme-runtime'
+import { buildPresenceScript } from '@/utils/achievements'
 import { type Metadata } from 'next'
 import { Noto_Sans, Space_Grotesk } from 'next/font/google'
