+++ src/components/layout/experience/work-history/work-history-item.tsx
@@ -1,6 +1,6 @@
-import type { WorkExperience } from '@/lib/experience'
-import { resolveSkills } from '@/lib/skillicons'
 import { formatMonthYear } from '@/lib/utils'
+import type { WorkExperience } from '@/types/work-experience'
+import { resolveSkills } from '@/utils/skillicons'
 import { CollapsibleHighlights } from './collapsible-highlights'
 import { CompanyMarker } from './company-marker'
