+++ src/components/providers/achievements-provider.tsx
@@ -1,16 +1,15 @@
 'use client'
 
+import { ACHIEVEMENTS, ACHIEVEMENTS_COOKIE_NAME, type AchievementId } from '@/lib/achievements'
+import type { ThemeName } from '@/lib/themes'
+import { cn } from '@/lib/utils'
 import {
-  ACHIEVEMENTS,
-  ACHIEVEMENTS_COOKIE_NAME,
   createEmptyUnlocked,
   parseUnlockedStorage,
   serializeUnlockedCookie,
-  type AchievementId,
   type UnlockedMap,
-} from '@/lib/achievements'
-import { getThemeBaseColor, type ThemeName } from '@/lib/themes'
-import { cn } from '@/lib/utils'
+} from '@/utils/achievements'
+import { getThemeBaseColor } from '@/utils/themes'
 import React, { createContext, useCallback, useContext, useEffect, useMemo, useRef, useState } from 'react'
 import { Toaster, toast } from 'sonner'
 import { useTheme } from './theme-provider'
