+++ src/lib/experience.ts
@@ -1,30 +1,7 @@
 import draftKingsLogo from '@/images/logos/draftkings.webp'
 import newInnovationsLogo from '@/images/logos/newinnov.webp'
 import qgendaLogo from '@/images/logos/qgenda.webp'
-import type { Route } from 'next'
-import type { StaticImageData } from 'next/image'
-import type { SkillName } from './skillicons'
-
-export interface WorkExperience {
-  id: string
-  role: string
-  company: string
-  location?: string
-  from: string
-  to?: string
-  summary: string
-  highlights?: string[]
-  skills?: SkillName[]
-  companyUrl?: Route
-  companyLogoSrc?: StaticImageData
-  companyLogoAlt?: string
-}
-
-export interface SkillCategory {
-  id: 'craft' | 'data' | 'platforms' | 'toolbox'
-  label: string
-  items: SkillName[]
-}
+import type { SkillCategory, WorkExperience } from '@/types/work-experience'
 
 export const WORK_HISTORY: WorkExperience[] = [
   {
