+++ src/lib/light-grid.ts
@@ -1,16 +1,4 @@
-export interface LightGrid {
-  GRID_SIZE_PX: number
-  GRID_OFFSET_PX: number
-  NUM_LIGHTS: number
-  DURATION_SECONDS: number
-  STAGGER_SECONDS: number
-  DETOUR_PROBABILITY: number
-  HORIZONTAL_PROBABILITY: number
-  DOT_SIZE_PX: number
-  COLOR_RGB: string
-  GLOW_NEAR_PX: number
-  GLOW_FAR_PX: number
-}
+import type { LightGrid } from '@/types/light-grid'
 
 export const LIGHT_GRID: LightGrid = {
   // layout
