+++ src/lib/projects.ts
@@ -3,22 +3,7 @@ import kilDevImage from '@/images/projects/kil-dev.webp'
 import kubernetesImage from '@/images/projects/kubernetes.webp'
 import questnestImage from '@/images/projects/questnest.webp'
 import tiaxslaughterImage from '@/images/projects/tiaxslaughter.webp'
-import type { Route } from 'next'
-import type { StaticImageData } from 'next/image'
-import type { SkillName } from './skillicons'
-
-export interface Project {
-  id: string
-  title: string
-  description: string
-  tags: SkillName[]
-  href?: Route
-  repo?: Route
-  year?: number
-  status?: 'live' | 'wip' | 'archived'
-  imageSrc: StaticImageData
-  imageAlt: string
-}
+import type { Project } from '@/types/projects'
 
 export const projects: Project[] = [
   {
