+++ src/types/home-content.d.ts
@@ -1,0 +1,7 @@
+export interface HomeContent {
+  NAME: string
+  TITLE: string
+  LOCATION: string
+  MAP_LATITUDE: number
+  MAP_LONGITUDE: number
+}
