+++ src/types/light-grid.d.ts
@@ -1,0 +1,13 @@
+export interface LightGrid {
+  GRID_SIZE_PX: number
+  GRID_OFFSET_PX: number
+  NUM_LIGHTS: number
+  DURATION_SECONDS: number
+  STAGGER_SECONDS: number
+  DETOUR_PROBABILITY: number
+  HORIZONTAL_PROBABILITY: number
+  DOT_SIZE_PX: number
+  COLOR_RGB: string
+  GLOW_NEAR_PX: number
+  GLOW_FAR_PX: number
+}
