+++ src/types/quick-facts.d.ts
@@ -1,0 +1,8 @@
+import type { Route } from 'next'
+
+export interface QuickFact {
+  label: string
+  value: string | React.ReactNode
+  href?: Route
+  note?: string
+}
