+++ src/types/social-links.d.ts
@@ -1,0 +1,6 @@
+import type { Route } from 'next'
+
+export interface SociallLinks {
+  GITHUB: Route
+  LINKEDIN: Route
+}
