+++ src/types/work-experience.d.ts
@@ -1,0 +1,24 @@
+import type { SkillName } from '@/lib/skillicons'
+import type { Route } from 'next'
+import type { StaticImageData } from 'next/image'
+
+export interface WorkExperience {
+  id: string
+  role: string
+  company: string
+  location?: string
+  from: string
+  to?: string
+  summary: string
+  highlights?: string[]
+  skills?: SkillName[]
+  companyUrl?: Route
+  companyLogoSrc?: StaticImageData
+  companyLogoAlt?: string
+}
+
+export interface SkillCategory {
+  id: 'craft' | 'data' | 'platforms' | 'toolbox'
+  label: string
+  items: SkillName[]
+}
